backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
redis==5.0.1
slowapi==0.1.9
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
scikit-learn==1.3.2
nltk==3.8.1
reportlab==4.0.8
jinja2==3.1.2
aiofiles==23.2.1
python-dateutil==2.8.2
